PoC/Model-Serving/vertex-predictor/requirements.txt
google-cloud-aiplatform[prediction]>=1.38.0
tensorflow-cpu>=2.11.1,<2.18.0
scikit-learn>=1.3.0
numpy>=1.24.3
